refacto(redux-store)
- moved all configuration related to redux-store in a separeted file `configureStore`
