+++ src/app/config/index.ts
@@ -1,3 +1,4 @@
 export * from './routes';
 export * from './Root';
 export * from './variables';
+export * from './configureStore';
