+++ src/app/custom-hooks/index.tsx
@@ -3,4 +3,3 @@ export * from './useForm';
 export * from './useToggler';
 export * from './useCountRenders';
 export * from './useSelector';
-export * from './useThunkDispatch';
