--- src/app/custom-hooks/useThunkDispatch.tsx
@@ -1,4 +1,0 @@
-import { useDispatch } from 'react-redux';
-import { store } from 'app/config';
-
-export const useThunkDispatch = () => useDispatch<typeof store.dispatch>();
